requirements.txt
# Web Framework and Async Stack
protobuf==5.29.5
fastapi>=0.95.0
uvicorn>=0.21.1
websockets>=11.0.3
requests>=2.28.2
pydantic>=1.10.7
aiohttp>=3.8.4

# Trading and Backtesting
backtrader>=1.9.78.123
yfinance>=0.2.18
ib_insync>=0.9.83
pandas>=1.5.3
numpy==1.26.4
matplotlib>=3.7.1

# Machine Learning and Optimization
scikit-learn>=1.2.2
optuna>=3.0.0


# TensorFlow (CPU) + Keras
tensorflow==2.14.0
keras==2.14.0

# ONNX and GPU runtime
onnx>=1.15
onnxruntime-gpu>=1.17.0

# LLM and Agent Framework
openai>=1.3.0
pyautogen>=0.2.0
autogenstudio>=0.1.0  
autogen-ext[openai]>=0.1.0

# Testing and Utilities
pytest>=7.3.1
python-multipart>=0.0.6
pylint
bandit
streamlit
# pandas_market_calendars==4.3 this not good see next line
pip install git+https://github.com/rsheftel/pandas_market_calendars.git
stable_baselines3
gym
shimmy>=2.0
talib
